requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
selenium==4.16.0
undetected-chromedriver==3.5.5
psycopg2-binary==2.9.9
tenacity==8.2.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
APScheduler==3.10.4
SQLAlchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
slack-sdk==3.23.0
